refactor: move type to own module
Also add type documentation where I knew what to write
